+++ prover/src/lib.rs
@@ -14,8 +14,8 @@ use miden_gpu::HashFn;
 use miden_processor::{
     ExecutionTrace, Program,
     crypto::{
-        Blake3_192, Blake3_256, ElementHasher, RandomCoin, Rpo256, RpoRandomCoin, Rpx256,
-        RpxRandomCoin, WinterRandomCoin,
+        Blake3_192, Blake3_256, ElementHasher, Poseidon2, RandomCoin, Rpo256, RpoRandomCoin,
+        Rpx256, RpxRandomCoin, WinterRandomCoin,
     },
     math::{Felt, FieldElement},
 };
@@ -124,6 +124,14 @@ pub fn prove(
             );
             #[cfg(all(feature = "metal", target_arch = "aarch64", target_os = "macos"))]
             let prover = gpu::metal::MetalExecutionProver::new(prover, HashFn::Rpx256);
+            maybe_await!(prover.prove(trace))
+        },
+        HashFunction::Poseidon2 => {
+            let prover = ExecutionProver::<Poseidon2, WinterRandomCoin<_>>::new(
+                options,
+                stack_inputs,
+                stack_outputs.clone(),
+            );
             maybe_await!(prover.prove(trace))
         },
     }